pyproject.toml
[project]
name = "skinbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.20.0.post0",
    "python-dotenv>=1.1.0",
    "ultralytics>=8.3.140",
]
